app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
simple-salesforce==1.12.6
pytest==8.3.3
httpx==0.27.2
openpyxl==3.1.5
python-multipart==0.0.20
